# apps/gateway/spec/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Play-Money Poker MVP API
  version: 0.1.0
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional idempotency key for safe retries of mutating requests.
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    UserProfile:
      type: object
      required: [userId, nickname, avatarUrl, stats, friends]
      properties:
        userId:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          nullable: true
        stats:
          type: object
          required: [handsPlayed, wins]
          properties:
            handsPlayed:
              type: integer
            wins:
              type: integer
        friends:
          type: array
          items:
            type: string
...
    FriendsList:
      type: object
      required: [friends]
      properties:
        friends:
          type: array
          items:
            type: string
...
    PushSubscription:
      type: object
      required: [endpoint]
      properties:
        endpoint:
          type: string
        keys:
          type: object
          properties:
            p256dh:
              type: string
            auth:
              type: string
    VapidPublicKey:
      type: object
      required: [publicKey]
      properties:
        publicKey:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/push/vapid:
    get:
      summary: Get VAPID public key
      responses:
        '200':
          description: VAPID public key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VapidPublicKey'
        '503':
          description: VAPID missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/push/subscribe:
    post:
      summary: Register push subscription
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushSubscription'
      responses:
        '204':
          description: Subscription stored
        '400':
          description: Invalid subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Remove push subscription
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint]
              properties:
                endpoint:
                  type: string
      responses:
        '204':
          description: Subscription removed
        '400':
          description: Invalid subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me:
    get:
      summary: Get current user profile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/profile:
    post:
      summary: Update nickname and avatar
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                avatarUrl:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/friends:
    get:
      summary: List friends
      responses:
        '200':
          description: Friends list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendsList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update friends list
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendsList'
      responses:
        '200':
          description: Updated friends list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendsList'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...