# apps/gateway/src/http/routes/push.ts
import type { Request, Response } from 'express';
import { Router } from 'express';
import { grpc } from '../../grpc/unaryClients';
import { safeAuthedRoute } from '../utils/safeAuthedRoute';
import { getIdempotencyKey } from '../utils/idempotencyKey';

const router = Router();

type ParsedSubscription =
  | { ok: true; subscription: { endpoint: string; keys: { p256dh: string; auth: string } } }
  | { ok: false; error: string };

function parseSubscription(body: unknown): ParsedSubscription {
  if (!body || typeof body !== 'object') {
    return { ok: false, error: 'Invalid subscription' };
  }

  const record = body as Record<string, unknown>;
  const endpoint = typeof record.endpoint === 'string' ? record.endpoint.trim() : '';
  if (!endpoint) {
    return { ok: false, error: 'endpoint is required' };
  }

  const keysValue = record.keys;
  if (!keysValue || typeof keysValue !== 'object') {
    return { ok: false, error: 'keys are required' };
  }

  const keysRecord = keysValue as Record<string, unknown>;
  const p256dh = typeof keysRecord.p256dh === 'string' ? keysRecord.p256dh.trim() : '';
  const auth = typeof keysRecord.auth === 'string' ? keysRecord.auth.trim() : '';
  if (!p256dh || !auth) {
    return { ok: false, error: 'keys.p256dh and keys.auth are required' };
  }

  return { ok: true, subscription: { endpoint, keys: { p256dh, auth } } };
}

router.get(
  '/vapid',
  safeAuthedRoute(
    async (_req: Request, res: Response) => {
      const publicKey = process.env.VAPID_PUBLIC_KEY;
      if (!publicKey) {
        res.status(503).json({ error: 'VAPID public key is not configured' });
        return;
      }

      res.json({ publicKey });
    },
    { logMessage: 'Failed to get VAPID public key' },
  ),
);

router.post(
  '/subscribe',
  safeAuthedRoute(
    async (req: Request, res: Response, userId: string) => {
      const parsed = parseSubscription(req.body);
      if (!parsed.ok) {
        res.status(400).json({ error: parsed.error });
        return;
      }

      const idempotencyKey = getIdempotencyKey(req);
      const response = await grpc.notify.RegisterSubscription({
        user_id: userId,
        subscription: parsed.subscription,
        idempotency_key: idempotencyKey,
      });

      if (!response.ok) {
        res.status(400).json({ error: response.error || 'Failed to register subscription' });
        return;
      }

      res.status(204).send();
    },
    { logMessage: 'Failed to register push subscription' },
  ),
);

router.delete(
  '/subscribe',
  safeAuthedRoute(
    async (req: Request, res: Response, userId: string) => {
      const endpoint = typeof req.body?.endpoint === 'string' ? req.body.endpoint.trim() : '';
      if (!endpoint) {
        res.status(400).json({ error: 'endpoint is required' });
        return;
      }

      const idempotencyKey = getIdempotencyKey(req);
      const response = await grpc.notify.UnregisterSubscription({
        user_id: userId,
        endpoint,
        idempotency_key: idempotencyKey,
      });

      if (!response.ok) {
        res.status(400).json({ error: response.error || 'Failed to unregister subscription' });
        return;
      }

      res.status(204).send();
    },
    { logMessage: 'Failed to unregister push subscription' },
  ),
);

export default router;